Move shared code to a shared package for sdk linking and make certain commands public so they are accessible.
This is to share code between convert.go and package_add.go
