+++ pkg/cmd/pulumi/packagecmd/package_extract_mapping.go
@@ -18,7 +18,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/cmd"
+	pcmd "github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/cmd"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/resource/plugin"
 	"github.com/spf13/cobra"
 )
@@ -33,7 +33,7 @@ func newExtractMappingCommand() *cobra.Command {
 		Long: `Get the mapping information for a given key from a package.
 
 <schema_source> can be a package name or the path to a plugin binary.`,
-		Run: cmd.RunCmdFunc(func(cmd *cobra.Command, args []string) error {
+		Run: pcmd.RunCmdFunc(func(cmd *cobra.Command, args []string) error {
 			key := args[0]
 			source := args[1]
 			var provider string
@@ -41,7 +41,7 @@ func newExtractMappingCommand() *cobra.Command {
 				provider = args[2]
 			}
 
-			p, err := providerFromSource(source)
+			p, err := pcmd.ProviderFromSource(source)
 			if err != nil {
 				return fmt.Errorf("load provider: %w", err)
 			}
