+++ pkg/go.mod
@@ -110,6 +110,7 @@ require (
 	golang.org/x/term v0.27.0
 	golang.org/x/text v0.21.0
 	google.golang.org/protobuf v1.35.1
+	gopkg.in/yaml.v2 v2.4.0
 	lukechampine.com/frand v1.4.2
 )
